Implemented Service Section

+++ Project/frontend/src/pages/manageservice/viewrep.js
@@ -1,19 +1,17 @@
 import React, { useEffect, useState } from "react";
 import { useParams, useNavigate } from "react-router-dom";
+import Swal from 'sweetalert2'
 import "./serdash.css"
 
 export default function ViewDel() {
 
     const [form, setForm] = useState({
         customerid: "",
-        employeeid: "",
         repairdate: "",
         itemname: "",
-        repaidescription: "",
-        repairfee: "",
+        description: "",
         imgurl: "",
-        rstatus: "",
-
+        status: "",
     });
     const params = useParams();
     const navigate = useNavigate();
@@ -44,40 +42,58 @@ export default function ViewDel() {
     }, [params.id, navigate]);
 
     async function deleteRecord(id) {
-        await fetch(`${process.env.REACT_APP_BACKEND_URL}/repair/delete/${id}`, {
-            method: "DELETE"
-        });
+        const swalWithBootstrapButtons = Swal.mixin({
+            customClass: {
+                confirmButton: 'text-white bg-gradient-to-r from-green-400 via-green-500 to-green-600 hover:bg-gradient-to-br focus:ring-4 focus:outline-none focus:ring-green-300 dark:focus:ring-green-800 font-medium rounded-lg text-sm px-5 py-2.5 text-center mr-2 mb-2',
+                cancelButton: 'text-white bg-gradient-to-r from-red-400 via-red-500 to-red-600 hover:bg-gradient-to-br focus:ring-4 focus:outline-none focus:ring-red-300 dark:focus:ring-red-800 font-medium rounded-lg text-sm px-5 py-2.5 text-center mr-2 mb-2'
+            },
+            buttonsStyling: false
+        })
 
-        navigate("/viewalldeliveries");
+        swalWithBootstrapButtons.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this! You will lose the Repair Details!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonText: 'Yes, delete it!',
+            cancelButtonText: 'No, cancel!',
+            reverseButtons: true
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`${process.env.REACT_APP_BACKEND_URL}/repair/delete/${id}`, {
+                    method: "DELETE"
+                });
+
+                navigate("/manageservice/viewallrep");
+                swalWithBootstrapButtons.fire(
+                    'Deleted!',
+                    'Repair Details has been deleted.',
+                    'success'
+                ).then((result) => {
+                    if (result.isConfirmed) {
+                        window.location.reload();
+                    }
+                })
+            } else if (
+                /* Read more about handling dismissals below */
+                result.dismiss === Swal.DismissReason.cancel
+            ) {
+                swalWithBootstrapButtons.fire(
+                    'Cancelled',
+                    'Repair Details Not Deleted:)',
+                    'info'
+                )
+            }
+        })
     }
 
-    // This method will delete a record
-    // async function deleteRecord(id) {
-    //     await fetch(`http://localhost:5000/${id}`, {
-    //         method: "DELETE"
-    //     });
-
-    //     const newRecords = records.filter((el) => el._id !== id);
-    //     setRecords(newRecords);
-    // }
-
-    // This method will map out the records on the table
-    // function recordList() {
-    //   return records.map((record) => {
-    //     return (
-    //       <ViewCusTag
-    //         record={record}
-    //         // deleteRecord={() => deleteRecord(record._id)}
-    //         key={record._id}
-    //       />
-    //     );
-    //   });
-    // }
     return (
         <div>
             <div className="viewcus dark:bg-gray-700 grid grid-cols-8 gap-4 hover:bg-gray-500 left">
                 <div className="col-span-3"><img src={form.imgurl} alt="" />
-                    <button type="button" class="text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 mr-2 mb-2 dark:bg-blue-600 dark:hover:bg-blue-700 focus:outline-none dark:focus:ring-blue-800">Update Details</button>
+                    <a href={`/manageservice/updaterep/${form._id}`}>
+                        <button type="button" class="text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 mr-2 mb-2 dark:bg-blue-600 dark:hover:bg-blue-700 focus:outline-none dark:focus:ring-blue-800">
+                            Update Details</button></a>
                     <button onClick={() => {
                         deleteRecord(form._id);
                     }}
@@ -85,18 +101,25 @@ export default function ViewDel() {
                     >Delete Details</button>
                 </div>
 
-                <div className="">
-                    <p className="text-lg sm">{form.itemname} </p>
-                    <p className="text-lg"><b>{form.customerid}</b></p>
-                    <br />
-                    <p className="text-lg"><b>{form.employeeid}</b></p>
-                    <p className="text-lg sm">{form.repairdate} </p>
-                    <p className="text-lg sm">{form.repaidescription} </p>
-                    <p className="text-lg sm">{form.repairfee} </p>
-                    <p className="text-lg sm">{form.imgurl} </p>
-                    <p className="text-lg sm">{form.rstatus} </p>
+                <div className="repitemcard">
+                    <div className="repitemcardtext">
+                        <ul>
+                            <li><p className="text-lg sm">Item Name - {form.itemname} </p></li>
+                            <li><p className="text-lg sm">Last Modification - {form.repairdate} </p></li>
+                            <li><p className="text-lg sm">Status - {form.status} </p></li>
+                            <li><p className="text-lg sm">Description <br /><br />{form.description} </p></li>
+                        </ul>
 
+                    </div>
+
+
+                    <a href={`/viewcus/${form.customerid}`}>
+                        <button type="button" class="text-gray-900 bg-white border border-gray-300 focus:outline-none hover:bg-gray-100 focus:ring-4 focus:ring-gray-200 font-medium rounded-lg text-sm px-5 py-2.5 mr-2 mb-2 dark:bg-gray-800 dark:text-white dark:border-gray-600 dark:hover:bg-gray-700 dark:hover:border-gray-600 dark:focus:ring-gray-700">
+                            View Customer Profile
+                        </button>
+                    </a>
                 </div>
+
 
 
             </div>
